pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["derijablya <dulugov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"
sqlalchemy = {extras = ["asynchio"], version = "^2.0.22"}
pydantic = {extras = ["email"], version = "^2.4.2"}
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"
black = "^23.10.1"
ruff = "^0.1.3"
pytest = "^7.4.3"
httpx = "^0.25.0"
asyncpg = "^0.28.0"
openpyxl = "^3.1.2"
pandas = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
